+++ src/00-namespace.js
@@ -1,4 +1,4 @@
 const BS = {};
 
-BS.VERSION = '1.0.3';
+BS.VERSION = '1.0.4';
 BS.TAG = '[bili-subtitle]';